[Setup] EUPL license + logo (#6)
This PR contains the work done to set the [European Union Public License](https://joinup.ec.europa.eu/collection/eupl/introduction-eupl-licence) license for this library, which also implies to update the header of the existing source code to reflect this license. Furthermore, a hand-drawn logo for this library has been included.

Co-authored-by: Javier Cicchelli <javier@rock-n-code.com>
Reviewed-on: https://repo.rock-n-code.com/rock-n-code/swift-libs/pulls/6